+++ packages/alpha/src/types.ts
@@ -16,6 +16,7 @@ import type {
 
 export * from '@cognite/sdk';
 export * from './api/workflows/types';
+export * from './api/dataProducts/types';
 
 // This file is here mostly to allow apis to import { ... } from '../../types';
 // Overriding types should probably be done in their respective API endpoint files, where possible