+++ packages/stable/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [10.12.1](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk@10.12.0...@cognite/sdk@10.12.1) (2026-06-23)
+
+**Note:** Version bump only for package @cognite/sdk
+
+
+
+
+
 # [10.12.0](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk@10.11.0...@cognite/sdk@10.12.0) (2026-06-05)